requirements.txt
Django~=4.2.0
djangorestframework
requests 